Add create button. All work, but page refreshing.

+++ src/main/java/it/academy/dto/request/StudentDTORequest.java
@@ -28,7 +28,7 @@ public class StudentDTORequest implements Serializable {
     private Integer age;
     @SerializedName("mark")
     @Expose
-    private Double mark;
+    private Integer mark;
     @SerializedName("city")
     @Expose
     private String city;
